% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trait-core.R
\name{bare_typed}
\alias{bare_typed}
\title{Check \code{typeof()}}
\usage{
bare_typed(type, typeof)
}
\arguments{
\item{type}{A type.}

\item{typeof}{A string. Must be a one of the valid R types returned by \code{\link[=typeof]{typeof()}}:
\code{"double"}, \code{"integer"}, \code{"logical"}, \code{"character"}, \code{"complex"},
\code{"raw"}, \code{"list"}, \code{"NULL"}, \code{"environment"}, \code{"symbol"}, \code{"pairlist"},
\code{"language"}, \code{"expression"}, \code{"S4"}, \code{"closure"}, \code{"special"},
\code{"builtin"}, \code{"externalptr"}, \code{"weakref"}, \code{"promise"}, \code{"char"}, or \code{"bytecode"}.}
}
\value{
A copy of \code{type} with an additional \code{typeof} constraint.
}
\description{
\code{bare_typed()} returns a copy of \code{type} that requires objects to be a bare
R value with \code{typeof()} equal to \code{typeof}. Bare objects are those with no
class attribute, for example \code{1L} or \code{list()}, but not \code{data.frame()}.

\if{html}{\out{<div class="sourceCode r">}}\preformatted{t_lgl <- t_any |> bare_typed("logical")
obj_is_type(1, t_lgl)  # FALSE
obj_is_type(NA, t_lgl) # TRUE
}\if{html}{\out{</div>}}
}
\examples{
t_pairlist <- t_any |> bare_typed("pairlist")
obj_inspect_type(10L, t_pairlist)
obj_inspect_type(pairlist(), t_pairlist)

}
