% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type-core.R
\name{base-types}
\alias{base-types}
\alias{t_any}
\alias{t_null}
\alias{t_list}
\alias{t_env}
\alias{t_fun}
\alias{t_vec}
\alias{t_num}
\alias{t_lgl}
\alias{t_bool}
\alias{t_int}
\alias{t_dbl}
\alias{t_chr}
\alias{t_string}
\alias{t_dataframe}
\alias{t_factor}
\alias{t_date}
\alias{t_posixct}
\alias{t_dots}
\title{Built-in types}
\usage{
t_any

t_null

t_list

t_env

t_fun

t_vec

t_num

t_lgl

t_bool

t_int

t_dbl

t_chr

t_string

t_dataframe

t_factor

t_date

t_posixct

t_dots
}
\description{
The \verb{t_*} types represent common objects in R, such as functions
and atomic vectors. They may be refined using traits, such as
\code{\link[=complete]{complete()}} or \code{\link[=bounded]{bounded()}}, to add additional restrictions to the
type.

Each \verb{t_*} type is an S7 object containing a list of traits. \code{t_bool}
for example is constructed from \code{\link[=bare_typed]{bare_typed()}}, \code{\link[=sized]{sized()}}, and
\code{\link[=complete]{complete()}}:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{t_bool <- bare_typed("logical") |> sized(1L) |> complete()
}\if{html}{\out{</div>}}

The following base types are provided:\tabular{ll}{
   Object \tab Matches \cr
   \code{t_any} \tab Any object \cr
   \code{t_null} \tab \code{NULL} \cr
   \code{t_list} \tab A list \cr
   \code{t_env} \tab An environment \cr
   \code{t_fun} \tab A function \cr
   \code{t_vec} \tab A \link[vctrs:vctrs-package]{vctrs}-style vector \cr
   \code{t_num} \tab An numeric (e.g. integer or double) vector \cr
   \code{t_lgl} \tab A bare logical vector \cr
   \code{t_bool} \tab A single \code{TRUE} or \code{FALSE} \cr
   \code{t_int} \tab A bare integer vector \cr
   \code{t_dbl} \tab A bare double vector \cr
   \code{t_chr} \tab A bare character vector \cr
   \code{t_string} \tab A single non-\code{NA} string \cr
   \code{t_dataframe} \tab A data frame \cr
   \code{t_factor} \tab A factor \cr
   \code{t_date} \tab A \code{Date} \cr
   \code{t_posixct} \tab A \code{POSIXct} datetime \cr
   \code{t_dots} \tab A \code{...} argument \cr
}
}
\examples{
obj_is_type(1L, t_int)
obj_is_type(1.5, t_int)

obj_is_type(TRUE, t_bool)
obj_is_type(NA, t_bool)
obj_is_type(c(TRUE, FALSE), t_bool)

obj_is_type("hello", t_string)
obj_is_type(NA_character_, t_string)

# Add traits to enforce additional restrictions
t_prob <- t_dbl |> bounded(0, 1)
t_name <- t_string |> within(c("x", "y", "z"))

}
\seealso{
\code{\link[=typed]{typed()}} for declaring typed functions.
}
