% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trait-core.R
\name{classed}
\alias{classed}
\title{Check \code{class()}}
\usage{
classed(type, classes, inherits = "all")
}
\arguments{
\item{type}{A type.}

\item{classes}{A character vector of class names.}

\item{inherits}{Whether the object must inherit from \code{"all"} supplied classes or at
least \code{"any"} one. Defaults to \code{"all"}.}
}
\value{
A copy of \code{type} with an additional class constraint.
}
\description{
\code{classed()} returns a copy of \code{type} that requires objects to inherit from one
or more classes. The \code{inherits} argument controls whether the object must
inherit from all supplied classes or at least one:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{# Must inherit from "Date"
t_date <- t_any |> classed("Date")

# Must inherit from both "POSIXct" and "POSIXt"
t_posixct <- t_any |> classed(c("POSIXct", "POSIXt"), inherits = "all")

# Must inherit from "Date" or "POSIXct"
t_datetime <- t_any |> classed(c("Date", "POSIXct"), inherits = "any")
}\if{html}{\out{</div>}}
}
\examples{
t_posixct <- t_any |> classed(c("POSIXct", "POSIXt"), inherits = "all")
obj_is_type(as.POSIXct("2020-01-01"), t_posixct)
obj_is_type(as.Date("2020-01-01"), t_posixct)

}
\seealso{
\code{\link[=bare_typed]{bare_typed()}} to constrain \code{typeof()} rather than \code{class()}.
}
