% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trait-core.R
\name{complete}
\alias{complete}
\title{Check for missing values}
\usage{
complete(type)
}
\arguments{
\item{type}{A type.}
}
\value{
A copy of \code{type} with an additional non-missingness constraint.
}
\description{
\code{complete()} returns a copy of \code{type} that requires objects to contain
no missing (i.e. \code{NA}) values, checked via \code{\link[vctrs:vec_any_missing]{vctrs::vec_any_missing()}}.
}
\examples{
t_real <- t_dbl |> complete()
obj_inspect_type(10.6, t_real)
obj_inspect_type(c(1, NA), t_real)

}
