% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trait-core.R
\name{contains}
\alias{contains}
\title{Check that object is a superset}
\usage{
contains(type, values)
}
\arguments{
\item{type}{A type.}

\item{values}{A non-empty, non-list vector of values that objects must contain.}
}
\value{
A copy of \code{type} with an additional constraint that objects contain
all elements of \code{values}.
}
\description{
\code{contains()} returns a copy of \code{type} that requires objects to contain all
elements of \code{values}, checked via \code{\link[vctrs:vec_in]{vctrs::vec_in()}}.
}
\examples{
t_rgb <- t_chr |> contains(c("r", "g", "b"))
obj_inspect_type(c("r", "g", "b", "a"), t_rgb)
obj_inspect_type(c("r", "g"), t_rgb)

}
