% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trait-core.R
\name{disjoint_to}
\alias{disjoint_to}
\title{Check that object does not contain values}
\usage{
disjoint_to(type, values)
}
\arguments{
\item{type}{A type.}

\item{values}{A non-empty, non-list vector of values that an object must not contain.}
}
\value{
A copy of \code{type} with an additional constraint that objects contain
none of \code{values}.
}
\description{
\code{disjoint_to()} returns a copy of \code{type} that requires an object to be
not contain any of \code{values}, checked via \code{\link[vctrs:vec_in]{vctrs::vec_in()}}.
}
\examples{
t_not_three <- t_int |> setequal_to(1:3)
obj_inspect_type(10L, t_not_three)
obj_inspect_type(c(1L, 2L, 3L, 1L), t_not_three)

}
