% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typed-object.R
\name{\%:\%}
\alias{\%:\%}
\title{Declare a typed object}
\usage{
type \%:\% declaration
}
\arguments{
\item{type}{A type, optionally wrapped in \code{\link[=const]{const()}}.}

\item{declaration}{A declaration of the form \code{name(value)}. \code{name} becomes
the variable name and \code{value} is used as the variable's initial
value.}
}
\value{
The initial value, invisibly.
}
\description{
\verb{\%:\%} declares a typed variable in the calling environment. The right-hand
side must be a declaration of the form \code{name(value)}, where \code{name} is the
variable to create and \code{value} is its initial value:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{t_int \%:\% x(1L)
x        # 1L
x <- 2L  # ok
x <- "A" # error, "A" is not an integer
}\if{html}{\out{</div>}}

A typed variable behaves like a regular object, but is type checked on
every assignment. A constant variable may be declared using \code{\link[=const]{const()}},
in which case the variable cannot be modified.

\if{html}{\out{<div class="sourceCode r">}}\preformatted{const(t_int) \%:\% x(1L)
x <- 2L # error, can't assign to a constant
}\if{html}{\out{</div>}}
}
\examples{
t_chr |> sized(1L) \%:\% x("hello")
const(t_bool) \%:\% y(TRUE)

# `x` must be a scalar character
x <- "goodbye"
try(x <- 10)

# `y` can't be assigned to
try(y <- FALSE)

# Typed variables can be re-declared as a different type
t_chr \%:\% y("true")

}
\seealso{
Modifier \code{\link[=const]{const()}}, \code{\link[=typed]{typed()}} for type-checking functions.
}
