% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trait-relations.R, R/type-aaa.R,
%   R/typed-object.R, R/typed.R
\name{inline_assert_same_classed}
\alias{inline_assert_same_classed}
\alias{inline_assert_same_sized}
\alias{inline_assert_recyclable}
\alias{inline_assert_exclusive}
\alias{inline_abort_mistyped}
\alias{inline_obj_type_validate}
\alias{inlined-functions}
\alias{inline_assert_type}
\alias{inline_arg_assert_type}
\alias{inline_dots_assert_type}
\alias{inline_dotlist_assert_type}
\title{Inlined helper functions}
\usage{
inline_assert_same_classed(..., error_call = rlang::caller_env())

inline_assert_same_sized(..., error_call = rlang::caller_env())

inline_assert_recyclable(..., error_call = rlang::caller_env())

inline_assert_exclusive(..., error_call = rlang::caller_env())

inline_abort_mistyped(
  type,
  message,
  error_subclass = character(),
  error_call = rlang::caller_env()
)

inline_obj_type_validate(obj, obj_name, type)

inline_assert_type(
  obj,
  obj_name,
  type,
  error_header,
  error_subclass = character(),
  error_call = rlang::caller_env()
)

inline_arg_assert_type(arg, type, arg_name, error_call = rlang::caller_env())

inline_dots_assert_type(..., .type, .error_call = rlang::caller_env())

inline_dotlist_assert_type(dots, type, error_call = rlang::caller_env())
}
\arguments{
\item{...}{Dots to be checked individually.}

\item{error_call, .error_call}{The call to use in error messages.}

\item{type, .type}{A type.}

\item{message}{An error message.}

\item{error_subclass}{An error class.}

\item{obj}{An object to be checked.}

\item{obj_name}{An object name to use in error messages.}

\item{error_header}{An error message header.}

\item{arg}{An argument to be checked.}

\item{arg_name}{An argument name to use in error messages.}

\item{dots}{A list of dots to be checked.}
}
\description{
These functions are inserted into \code{\link[=typed]{typed()}} functions and are not meant
for external use. They are exported only to ensure that \code{\link[=typed]{typed()}} functions
call the correct helper, e.g. \code{type::inline_assert_type()}, and
not a globally defined function, e.g. \code{inline_assert_type()}.
}
\examples{
try(inline_dots_assert_type(10L, t_chr))
}
