% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type-aaa.R
\name{is_type}
\alias{is_type}
\title{Test if an object is a type}
\usage{
is_type(x)
}
\arguments{
\item{x}{An object to test.}
}
\value{
\code{TRUE} if \code{x} is a type, \code{FALSE} otherwise.
}
\description{
Test if an object is a type
}
\examples{
is_type(10L)
is_type(t_int)

}
