% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type-aaa.R
\name{last_type}
\alias{last_type}
\title{Return the expected type of the last mistyped object}
\usage{
last_type()
}
\value{
The last expected type. If no type assertions have been run, returns \code{NULL}.
}
\description{
\code{last_type()} returns the expected type of the last object to
fail a type check in \code{\link[=obj_assert_type]{obj_assert_type()}} or \code{\link[=typed]{typed()}}.
}
\examples{
# Returns `NULL` if no type checks have been run
last_type()

# `last_type()` returns `t_bool` after failed assertion
\dontrun{
obj_assert_type(10L, t_bool)
last_type()
}

}
