% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modifications.R
\name{modifiers}
\alias{modifiers}
\alias{const}
\alias{optional}
\alias{maybe}
\title{Modify a type}
\usage{
const(type)

optional(type)

maybe(type)
}
\arguments{
\item{type}{A type, e.g. \link{t_int}.}
}
\value{
A modified copy of \code{type}.
}
\description{
Type modifiers adjust how a type is enforced in a given context:
\itemize{
\item \code{const(type)}: For use in \verb{\%:\%}. Declares a constant.
\item \code{optional(type)}: For use in \code{\link[=typed]{typed()}}. Allows an argument to be missing.
\item \code{maybe(type)}: For use in \code{\link[=typed]{typed()}}. Allows an argument to be \code{NULL}.
}
}
\examples{
# const() variables can't be assigned to
const(t_int) \%:\% x(1L)
x
try(x <- 2L)

# optional() arguments may be missing
f <- typed(function(x = optional(t_int)) {
  if (missing(x)) { "absent" } else { x }
})
f()
f(1L)

# maybe() arguments may be `NULL`
g <- typed(function(x = maybe(t_int)) x)
g(NULL)
g(1L)

}
