% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type-aaa.R
\name{obj-type}
\alias{obj-type}
\alias{obj_is_type}
\alias{obj_inspect_type}
\alias{obj_assert_type}
\title{Check an object against a type}
\usage{
obj_is_type(obj, type)

obj_inspect_type(obj, type, obj_name = rlang::caller_arg(obj))

obj_assert_type(obj, type, obj_name = rlang::caller_arg(obj))
}
\arguments{
\item{obj}{An object to check.}

\item{type}{A type to check against.}

\item{obj_name}{The name of \code{obj}, used in messages. Defaults to the expression passed to \code{obj}.}
}
\value{
\itemize{
\item \code{obj_is_type()}: \code{TRUE} if \code{obj} is the correct type, \code{FALSE} otherwise.
\item \code{obj_inspect_type()}: \code{NULL} invisibly, called for its side effect.
\item \code{obj_assert_type()}: \code{NULL} invisibly if \code{obj} is the correct type, otherwise raises an error.
}
}
\description{
These functions are used to investiage whether an object \code{obj} has a type \code{type}.
\itemize{
\item \code{obj_is_type(obj, type)} returns \code{TRUE} if \code{obj} has type \code{type} and \code{FALSE} otherwise.
\item \code{obj_inspect_type(obj, type)} prints a success or failure message for each type check run on \code{obj} then returns \code{NULL} invisibly.
\item \code{obj_assert_type(obj, type)} raises a \verb{<type_error_mistyped_obj>} error if \code{obj} is mistyped and returns \code{NULL} invisibly otherwise.
}
}
\examples{
good <- TRUE
bad <- NA

# Test whether an object is a boolean
obj_is_type(good, t_bool)
obj_is_type(bad, t_bool)

# Print the type tests run on the object
obj_inspect_type(good, t_bool)
obj_inspect_type(bad, t_bool)

# Raise an error if the object is not a boolean
obj_assert_type(good, t_bool)
try(obj_assert_type(bad, t_bool))

}
