% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selector.R
\name{on}
\alias{on}
\alias{on_attr}
\alias{on_elm}
\alias{on_data}
\alias{on_each}
\alias{on_elms}
\alias{on_attrs}
\title{Select an element or attribute to type}
\usage{
on(accessor)

on_attr(name)

on_elm(index)

on_data()

on_each()

on_elms(indices)

on_attrs(attrs)
}
\arguments{
\item{accessor}{A call using \code{.x} as the object placeholder, or a bare symbol \code{f} as
shorthand for \code{f(.x)}.}

\item{name}{For \code{on_attr()}, a single attribute name (character).}

\item{index}{For \code{on_elm()}, a single position (integer) or name (character).}

\item{indices}{For \code{on_elms()}, positions (integer) or names (character).}

\item{attrs}{For \code{on_attrs()}, attribute names (character).}
}
\value{
A selector. These functions may only be used within \code{\link[=has]{has()}} and relations
such as \code{\link[=same_sized]{same_sized()}} and \code{\link[=same_classed]{same_classed()}}. Outside of these contexts, the
\code{on()} functions raise an error.
}
\description{
Selectors identify a part of an object for use in \code{\link[=has]{has()}} and relation
functions such as \code{\link[=same_sized]{same_sized()}} and \code{\link[=same_classed]{same_classed()}}.

These are useful for typing elements of a container type, such as a list. For
example, the following defines a coordinate type \code{t_coords}, which is a list
containing named numeric elements \code{"lat"} and \code{"lon"} of the same size:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{t_coords <- t_list |>
  has(on_elm("lat"), t_dbl |> bounded(-90, 90)) |>
  has(on_elm("lon"), t_dbl |> bounded(-180, 180)) |>
  has_relation(same_sized(on_elm("lat"), on_elm("lon")))
}\if{html}{\out{</div>}}

\code{on()} specifies an \code{accessor} function, either a call or a function name,
applied to an object during type checking. If \code{accessor} is a call, \code{.x}
may be used as a placeholder for the object.

\if{html}{\out{<div class="sourceCode r">}}\preformatted{# Require an object's length to be between 1 and 10 (inclusive)
t_any |> has(on(length(.x)), t_int |> bounded(1L, 10L))

# Require an object's names to contain "x" and "y"
# Equivilant to `on(names(.x))`
t_any |> has(on(names), t_chr |> contains(c("x", "y")))
}\if{html}{\out{</div>}}

\code{on_elm(index)} and \code{on_attr(name)} are convenience selectors for the
common cases of \code{on(.x[[index]])} and \code{on(attr(.x, name))}.

\if{html}{\out{<div class="sourceCode r">}}\preformatted{has(t_any, on_elm(1L), t_int)     # same as `on(.x[[1L]])`
has(t_any, on_attr("dim"), t_int) # same as `on(attr(.x, "dim"))`
}\if{html}{\out{</div>}}

\code{on_elms(indices)} and \code{on_attrs(names)} are the multiple selector forms
of \code{on_elm()} and \code{on_attr()}, useful for selecting multiple elements
in a relation:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{# These are equivilant
t_any |> has_relation(same_sized(on_elms(c(1L, 2L))))
t_any |> has_relation(same_sized(on_elm(1L), on_elm(2L)))
}\if{html}{\out{</div>}}

\code{on_data()} selects the underlying data of an object after it's attributes
and class has been removed (via \code{\link[=unclass]{unclass()}}).

\if{html}{\out{<div class="sourceCode r">}}\preformatted{# POSIXct datetime vectors store time as a double
t_posixct <- t_any |>
  classed("POSIXct") |>
  has(on_data(), t_dbl) |>
  has(on_attr("tzone"), t_chr |> sized(1L))
}\if{html}{\out{</div>}}

\code{on_each()} selects all elements of an object, applying the type check to
each one individually.

\if{html}{\out{<div class="sourceCode r">}}\preformatted{# Defines a list-of-integers type
t_list_of_int <- t_list |> has(on_each(), t_int)
}\if{html}{\out{</div>}}
}
\examples{
t_coords <- t_list |>
  has(on(names), t_chr |> setequal_to(c("lat", "lon"))) |>
  has(on_elm("lat"), t_dbl |> bounded(-90, 90)) |>
  has(on_elm("lon"), t_dbl |> bounded(-180, 180)) |>
  has_relation(same_sized(on_elms(c("lat", "lon"))))

good <- list(lat = c(70, 20, -50), lon = c(100, 0, 85))
obj_is_type(good, t_coords)

bad <- list(lat = c(1, 7), lon = 90)
obj_inspect_type(bad, t_coords)

}
\seealso{
\code{\link[=has]{has()}} to attach a selector to a type, \code{\link[=has_relation]{has_relation()}} to attatch a relation to a type.
}
