% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trait-core.R
\name{same_as}
\alias{same_as}
\title{Check that object is identical to a vector}
\usage{
same_as(type, values)
}
\arguments{
\item{type}{A type.}

\item{values}{A non-empty, non-list vector that an object must be identical to.}
}
\value{
A copy of \code{type} with an additional constraint that elements of objects
are identical to \code{values}.
}
\description{
\code{same_as()} returns a copy of \code{type} that requires an object to be
identical to \code{values}, checked via \code{\link[vctrs:vec_equal]{vctrs::vec_equal()}}.
}
\examples{
t_abc <- t_chr |> same_as(c("a", "b", "c"))
obj_inspect_type(c("a", "b", "c"), t_abc)
obj_inspect_type(c("c", "b", "a"), t_abc)

}
\seealso{
\code{\link[=setequal_to]{setequal_to()}} to test for equality, ignoring duplicates.
}
