% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trait-relations.R
\name{same_classed}
\alias{same_classed}
\title{Require elements to have the same class}
\usage{
same_classed(...)
}
\arguments{
\item{...}{Inside \code{\link[=has_relation]{has_relation()}}, one or more selectors (e.g. \code{\link[=on_elm]{on_elm()}},
\code{\link[=on_elms]{on_elms()}}, \code{\link[=on_attr]{on_attr()}}, \code{\link[=on]{on()}}).

Inside \code{\link[=typed]{typed()}}, one or more argument names.}
}
\description{
\code{same_classed()} constrains a set of values to share the same class. Its
behaviour depends on context:
\itemize{
\item Inside \code{\link[=has_relation]{has_relation()}}: requires that the selected elements or attributes of an object all have the same class.
\item Inside \code{\link[=typed]{typed()}}: requires that the specified function arguments all have the same class at call time.
}
}
\examples{
# Require that elements `[[1]]` and `[[2]]` share a class
t <- t_any |> has_relation(same_classed(on_elm(1L), on_elm(2L)))
obj_inspect_type(list(1L, 2L), t) 
obj_inspect_type(list(1L, "a"), t)

# Require that arguments `x` and `y` share a class
f <- typed(same_classed(x, y), function(x, y) { NULL })
f(1L, 2L)
try(f(1L, "a"))

}
\seealso{
\code{\link[=typed]{typed()}} for typed function construction, \code{\link[=has_relation]{has_relation()}} for
adding between-element type constraints.
}
