% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trait-core.R
\name{setequal_to}
\alias{setequal_to}
\title{Check that object is a given set}
\usage{
setequal_to(type, values)
}
\arguments{
\item{type}{A type.}

\item{values}{A non-empty, non-list vector of values that an object must be setequal to.}
}
\value{
A copy of \code{type} with an additional constraint that objects are setequal to \code{values}.
}
\description{
\code{setequal_to()} returns a copy of \code{type} that requires an object to be
setequal to \code{values}, checked via \code{\link[vctrs:vec_in]{vctrs::vec_in()}}.
}
\examples{
t_three <- t_int |> setequal_to(1:3)
obj_inspect_type(c(1L, 2L, 3L, 1L), t_three)
obj_inspect_type(10L, t_three)

}
