% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trait-core.R
\name{sized}
\alias{sized}
\title{Constrain the size of a type}
\usage{
sized(type, size)
}
\arguments{
\item{type}{A type.}

\item{size}{A non-negative count.}
}
\value{
A copy of \code{type} with an additional size constraint.
}
\description{
\code{sized()} returns a copy of \code{type} that requires objects to be
size \code{size}, checked via \code{\link[vctrs:vec_size]{vctrs::vec_size()}}.

\if{html}{\out{<div class="sourceCode r">}}\preformatted{obj_is_type(1, t_any |> sized(2L))   # FALSE
obj_is_type(1:2, t_any |> sized(2L)) # TRUE
}\if{html}{\out{</div>}}
}
\examples{
t_int2 <- t_int |> sized(2L)
obj_inspect_type(10L, t_int2)
obj_inspect_type(1:2, t_int2)

}
