% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type-package.R
\docType{package}
\name{type-package}
\alias{type}
\alias{type-package}
\title{type: Runtime Type Validation for Functions and Objects}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides a lightweight type system for 'R'. Types are represented as composable sets of predicates which a valid object must satisfy. Variables may be declared with a type, which is checked everytime the variable is assigned to. Additionally, function arguments and return values may be annotated with a type, which is validated each time the function is called.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/EthanSansom/type}
  \item \url{https://ethansansom.github.io/type/}
  \item Report bugs at \url{https://github.com/EthanSansom/type/issues}
}

}
\author{
\strong{Maintainer}: Ethan Sansom \email{ethan.sansom29@gmail.com} (\href{https://orcid.org/0009-0000-1573-0186}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Ethan Sansom \email{ethan.sansom29@gmail.com} (\href{https://orcid.org/0009-0000-1573-0186}{ORCID}) [copyright holder]
}

}
\keyword{internal}
