% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type-aaa.R
\name{type_union}
\alias{type_union}
\title{Declare a type union}
\usage{
type_union(...)
}
\arguments{
\item{...}{Two or more types to combine. At least one must be supplied.}
}
\value{
A type that accepts objects matching any of the supplied types.
}
\description{
\code{type_union()} returns a type that requires an object to satisfy at least
one of the types supplied to \code{...}.

\if{html}{\out{<div class="sourceCode r">}}\preformatted{# Must be an integer OR a double
t_number <- type_union(t_int, t_dbl)

# Must be a string OR NULL
t_opt_string <- type_union(t_string, t_null)
}\if{html}{\out{</div>}}

Nested unions are flattened, so \code{type_union(type_union(t_int, t_dbl), t_chr)}
is equivalent to \code{type_union(t_int, t_dbl, t_chr)}. Duplicate types are
silently dropped.
}
\examples{
t_index <- type_union(t_int, t_chr)
obj_is_type(1L, t_index)
obj_is_type("a", t_index)
obj_is_type(10.6, t_index)

# Useful for nullable types
t_opt_chr <- type_union(t_chr, t_null)
obj_is_type(c("a", "b"), t_opt_chr)
obj_is_type(NULL, t_opt_chr)
obj_is_type(1L, t_opt_chr)

}
\seealso{
\code{\link[=is_type]{is_type()}}, \code{\link[=is_type_union]{is_type_union()}} to test whether an object is a any type a or type union.
}
