% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typed.R
\name{typed}
\alias{typed}
\title{Declare a type-checked function}
\usage{
typed(..., returns = NULL)
}
\arguments{
\item{...}{A type annotated function definition, optionally accompanied by one or more
relation calls (e.g. \code{\link[=same_sized]{same_sized()}}). Exactly one function definition must
be supplied.}

\item{returns}{A type for the function's return value. By default \code{returns} is \code{NULL},
meaning the return value can have any type.}
}
\value{
A typed function.
}
\description{
\code{typed()} inserts argument and (optionally) return type validation into
a function. When printed, a typed function shows the types of its arguments
and return value.
}
\section{Argument Typing}{


A typed function's arguments may be annotated with any valid type, e.g.
\link{t_int}, including types refined with additional traits, e.g. \code{\link[=sized]{sized()}}.

\if{html}{\out{<div class="sourceCode r">}}\preformatted{f <- typed(function(x = t_int, y = t_chr |> sized(1L)) \{ paste(x, y) \})
f(1L, "a")         # ok
f(TRUE, "a")       # error, `x` is not an integer
f(1L, c("a", "b")) # error, `y` is not a size 1
}\if{html}{\out{</div>}}

By default, annotated function arguments have no default value. To set one,
use the syntax \verb{arg = <type> \%:\% <default>}.

\if{html}{\out{<div class="sourceCode r">}}\preformatted{f <- typed(function(x = t_int \%:\% 0L) \{ x \})
f()    # 0L
f(1L)  # 1L
}\if{html}{\out{</div>}}

Dots (\code{...}) may also be type annotated. By default, each dot will be checked
against the supplied type. To treat the dots as a single argument, e.g. as
a list, use the \link{t_dots} annotation.

\if{html}{\out{<div class="sourceCode r">}}\preformatted{# Each `...` must be an integer
f <- typed(function(... = t_int) list(...))

# Exactly 2 dots must be supplied
g <- typed(function(... = t_dots |> sized(2L)) list(...))
}\if{html}{\out{</div>}}

Arguments, with the exception of \code{...}, may be modified using \code{\link[=optional]{optional()}}
or \code{\link[=maybe]{maybe()}}. \code{\link[=optional]{optional()}} arguments may by unsupplied while \code{\link[=maybe]{maybe()}}
arguments may be \code{NULL}.

\if{html}{\out{<div class="sourceCode r">}}\preformatted{# `x` is an integer or is unsupplied
f <- typed(function(x = optional(t_int)) if (missing(x)) 0L else x)

# `x` is an integer or `NULL`
g <- typed(function(x = maybe(t_int)) x)
}\if{html}{\out{</div>}}
}

\section{Relations}{


Relations, e.g. \code{\link[=same_sized]{same_sized()}}, declare between-argument checks. A relation
call may be placed before or after the function definition.

\if{html}{\out{<div class="sourceCode r">}}\preformatted{# `x` and `y` must be integers of the same size
f <- typed(
  same_sized(x, y), 
  function(x = t_int, y = t_int) x + y
)
}\if{html}{\out{</div>}}
}

\section{Return Typing}{


Use the \code{returns} argument to enforce a type on the return value. Return
types cannot be modified using \code{\link[=optional]{optional()}} or \code{\link[=maybe]{maybe()}}.

\if{html}{\out{<div class="sourceCode r">}}\preformatted{f <- typed(
  function(x = t_bool) if (x) "yes" else "no", 
  returns = t_chr
)
}\if{html}{\out{</div>}}
}

\examples{
any2 <- typed(
  function(... = t_lgl, na.rm = t_bool \%:\% FALSE) { 
    any(..., na.rm) 
  },
  returns = t_lgl |> sized(1L)
)
print(any2)

# Correctly typed inputs proceed normally
any2(c(TRUE, FALSE), TRUE, na.rm = FALSE)

# Mistyped inputs cause an error
try(any2(TRUE, na.rm = "no"))

}
\seealso{
\code{\link[=untyped]{untyped()}} for un-typing a function, \code{\link[=optional]{optional()}} and \code{\link[=maybe]{maybe()}} for argument modifications, \code{\link[=same_sized]{same_sized()}} and \code{\link[=same_classed]{same_classed()}} for between-argument constraints.
}
