% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trait-core.R
\name{unduplicated}
\alias{unduplicated}
\title{Check for duplicate values}
\usage{
unduplicated(type)
}
\arguments{
\item{type}{A type.}
}
\value{
A copy of \code{type} with an additional non-duplicate constraint.
}
\description{
\code{unduplicated()} returns a copy of \code{type} that requires objects to contain
no duplicate values, checked via \code{\link[vctrs:vec_duplicate_any]{vctrs::vec_duplicate_any()}}.
}
\examples{
t_ids <- t_chr |> unduplicated()
obj_inspect_type(c("a1", "a2"), t_ids)
obj_inspect_type(c("a1", "b1", "b1"), t_ids)

}
