% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typed.R
\name{untyped}
\alias{untyped}
\title{Remove argument and return type validation from a typed function}
\usage{
untyped(fun)
}
\arguments{
\item{fun}{A function to untype.}
}
\value{
A function, without \code{\link[=typed]{typed()}} arguments or return values.
}
\description{
\code{untyped()} removes argument and return types from a typed function declared
via \code{\link[=typed]{typed()}}. If \code{fun} is a non-typed function, then \code{fun} is returned
as is.
}
\examples{
foo <- typed(function(x = t_int) { x })
print(foo)
print(untyped(foo))

}
