% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trait-core.R
\name{within}
\alias{within}
\title{Check that object is a subset}
\usage{
within(type, values)
}
\arguments{
\item{type}{A type.}

\item{values}{A non-empty, non-list vector of values that elements of object must be within.}
}
\value{
A copy of \code{type} with an additional constraint that objects are within \code{values}.
}
\description{
\code{within()} returns a copy of \code{type} that requires every element of an object
to be within \code{values}, checked via \code{\link[vctrs:vec_in]{vctrs::vec_in()}}.
}
\examples{
t_weekend <- t_chr |> within(c("Sat", "Sun"))
obj_inspect_type("Sat", t_weekend)
obj_inspect_type(c("A", "B"), t_weekend)

}
