% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trait-has.R
\name{has}
\alias{has}
\title{Add an element or attribute constraint to a type}
\usage{
has(type, selector, on_type)
}
\arguments{
\item{type}{A type.}

\item{selector}{A selector, e.g. the result of \code{\link[=on]{on()}}, \code{\link[=on_elm]{on_elm()}}, or \code{\link[=on_attr]{on_attr()}}.}

\item{on_type}{A type to check the selected value against.}
}
\value{
A copy of \code{type} with an additional element or attributeconstraint.
}
\description{
\code{has()} returns a copy of \code{type} that requires a selected part of an object
to have type \code{on_type}. The part is identified by a selector (see \code{\link[=on]{on()}}).

\if{html}{\out{<div class="sourceCode r">}}\preformatted{# Require that element `[[1]]` is an integer
t_any |> has(on_elm(1L), t_int)

# Require that the "dim" attribute is an integer vector of size 2
t_any |> has(on_attr("dim"), t_int |> sized(2L))

# Require that names contains "x"
t_any |> has(on(names), t_chr |> contains("x"))
}\if{html}{\out{</div>}}

Constraints can be composed with the pipe operator \verb{|>}:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{t_coords <- t_list |>
  has(on_elm("lat"), t_dbl |> bounded(-90, 90)) |>
  has(on_elm("lon"), t_dbl |> bounded(-180, 180))
}\if{html}{\out{</div>}}
}
\examples{
# TODO

}
\seealso{
\code{\link[=on]{on()}} for available selectors, \code{\link[=has_relation]{has_relation()}} to add between-element constraints.
}
