% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trait-relations.R
\name{has_relation}
\alias{has_relation}
\title{Add a between-element constraint to a type}
\usage{
has_relation(type, relation)
}
\arguments{
\item{type}{A type.}

\item{relation}{A relation, e.g. the result of \code{\link[=same_sized]{same_sized()}} or \code{\link[=same_classed]{same_classed()}}.}
}
\value{
A copy of \code{type} with an additional between-element constraint.
}
\description{
\code{has_relation()} returns a copy of \code{type} that requires a relationship to
hold between selected parts of an object. Parts are selected using selector
functions (see \link{on}) and the relationship is expressed as a relation
(see \code{\link[=same_sized]{same_sized()}}, \code{\link[=same_classed]{same_classed()}}).

\if{html}{\out{<div class="sourceCode r">}}\preformatted{# Require that elements `[[1]]` and `[[2]]` are the same size
t_any |> has_relation(same_sized(on_elm(1L), on_elm(2L)))

# Require that attributes "x" and "y" have the same class
t_any |> has_relation(same_classed(on_attr("x"), on_attr("y")))
}\if{html}{\out{</div>}}

\code{has()} and \code{has_relation()} can be composed:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{t_coords <- t_list |>
  has(on_elm("lat"), t_dbl |> bounded(-90, 90)) |>
  has(on_elm("lon"), t_dbl |> bounded(-180, 180)) |>
  has_relation(same_sized(on_elm("lat"), on_elm("lon")))
}\if{html}{\out{</div>}}
}
\examples{
# TODO

}
\seealso{
\code{\link[=same_sized]{same_sized()}} and \code{\link[=same_classed]{same_classed()}} for available relations, \code{\link[=has]{has()}} to add per-element constraints.
}
